# Disable luindex (and other related things)