*Needs testing before merging*

This is three things in one, so it's best to read the individual commits.

This needs at least light testing before merging. Edd, can you look into that please? I'd suggest reducing `ITERATIONS` to (say) 10 and `PROCESSES` to (say) 3, which shouldn't take *too* long to execute.