@ltratt 2017-03-21
OK, see if that squashing works for you.