@vext01 2017-03-21
Ah, one more detail, which i need to test actually, I think you need:

```diff
                     stdout, stderr, rc = run_shell_cmd_bench(
                         "%s -jar %s %s -n %s" % (jvm_cmd, JAR, benchmark,
-                                                 ITERATIONS + 1), platform)
+                                                 ITERATIONS + 1), platform, failure_fatal=False)
```

or `run_shell_cmd_bench` is gong to stop the runner.