@ltratt 2017-03-21
Shall I do that and squash at the same time or ...?