@vext01 2017-03-21
I can test this by killing one thread of the vm during a run.

Before:

```
  File "/home/kruninit/warmup_experiment/krun/krun/util.py", line 52, in fatal
    raise FatalKrunError(msg)
krun.util.FatalKrunError: Command failed: '/home/kruninit/warmup_experiment/work/mx/mx --java-home=/home/kruninit/warmup_experiment/work/graal-jvmci-8/jdk1.8.0/product -p /home/kruninit/warmup_experiment/work/graal/ vm -XX:+UseJVMCICompiler -jar extbench/dacapo-9.12-bach.jar avrora -n 11'
stdout:

stderr:
===== DaCapo 9.12 avrora starting warmup 1 =====
===== DaCapo 9.12 avrora completed warmup 1 in 3078344857 nsec =====
...
```

After:
```
graal:
  avrora: 0
===== DaCapo 9.12 avrora starting warmup 1 =====
===== DaCapo 9.12 avrora completed warmup 1 in 3047048276 nsec =====
===== DaCapo 9.12 avrora starting warmup 2 =====
===== DaCapo 9.12 avrora completed warmup 2 in 2060842600 nsec =====
===== DaCapo 9.12 avrora starting warmup 3 =====
===== DaCapo 9.12 avrora completed warmup 3 in 1953213189 nsec =====
===== DaCapo 9.12 avrora starting warmup 4 =====
===== DaCapo 9.12 avrora completed warmup 4 in 1932656164 nsec =====
===== DaCapo 9.12 avrora starting warmup 5 =====
===== DaCapo 9.12 avrora completed warmup 5 in 1920809910 nsec =====
===== DaCapo 9.12 avrora starting warmup 6 =====

 1 2
```

Can you add some kind of a message that indicates that the pexec failed. Otherwise I think that has worked OK.

In the latter case, you will get a results file like:
```
processnum,benchmark,0,1,2,3,4,5,6,7,8,9
1,avrora,3.07552784,2.068373501,1.941320083,1.936208724,1.93754041,1.930098086,1.913583866,1.93886327,1.909749109,1.924084161
2,avrora,3.081634177,2.061214975,1.923640755,1.945695335,1.927211131,1.930267844,1.931112786,1.923637596,1.902661694,1.89429299
```

i.e. with a missing line. You OK with that? I think perhaps we might need to insert some kind of a crash marker, or adjust the consumers so-as they don't blindly assume the lines are consecutive pexecs. If you see what I mean?