@vext01 2017-03-21
From a quick look at the conversion code, it looks like the pexec number is ignored, and all pexecs are assumed to be consecutive:
https://github.com/softdevteam/warmup_experiment/blob/master/warmup/krun_results.py#L35

So if we have pexecs `[1,3,7,9]` in our CSV, we would have pexecs `[0,1,2,3]` once converted. IIRC, the convention used in krun for a crashed pexec is the empty list, but let me check.