@ltratt 2017-03-21
Wouldn't the important code be the `csv_to_krun` code in this case?