@vext01 2017-03-21
It imports and uses the code i linked above:
https://github.com/softdevteam/warmup_experiment/blob/master/bin/csv_to_krun_json#L17