@snim2 2017-03-21
Probably sensible to just raise an issue on this, is there a test case yet?