@vext01 2017-03-21
We can do a separate PR yeah. Sarah and I can handle the crashing case.

I'm worried that having the wrong number of elements in the results file will confuse other parts of our code, so I'd like to look into it quickly.