@ltratt 2017-03-21
I still need to add `failure_fatal=False` I assume?